--- 3o14/src/federation.ts
@@ -1,356 +1,0 @@
-import {
-  Accept,
-  type Actor as APActor,
-  Create,
-  createFederation,
-  Endpoints,
-  exportJwk,
-  Follow,
-  generateCryptoKeyPair,
-  getActorHandle,
-  importJwk,
-  isActor,
-  Note,
-  Person,
-  PUBLIC_COLLECTION,
-  type Recipient,
-  Undo,
-} from "@fedify/fedify";
-import { getLogger } from "@logtape/logtape";
-import { InProcessMessageQueue, MemoryKvStore } from "@fedify/fedify";
-import { Temporal } from "@js-temporal/polyfill";
-import db from "./db.ts";
-import type { Actor, Post, User, Key } from "./models/index.ts";
-
-const logger = getLogger("3o14");
-
-const federation = createFederation({
-  kv: new MemoryKvStore(),
-  queue: new InProcessMessageQueue(),
-});
-
-// return the user with specified username if available
-// on this instance
-federation
-  .setActorDispatcher(
-    "/users/{identifier}",
-    async (ctx, identifier) => {
-      const user = db
-        .prepare<unknown[], User & Actor>(
-          `
-          SELECT * FROM users
-          JOIN actors ON (users.id = actors.user_id)
-          WHERE users.username = ?
-        `,
-        )
-        .get(identifier);
-      if (user == null) return null;
-
-      const keys = await ctx.getActorKeyPairs(identifier);
-      return new Person({
-        id: ctx.getActorUri(identifier),
-        preferredUsername: identifier,
-        name: user.name,
-        inbox: ctx.getInboxUri(identifier),
-        endpoints: new Endpoints({
-          sharedInbox: ctx.getInboxUri(),
-        }),
-        url: ctx.getActorUri(identifier),
-        publicKey: keys[0].cryptographicKey,
-        assertionMethods: keys.map((k) => k.multikey),
-        followers: ctx.getFollowersUri(identifier),
-      });
-    },
-  )
-  .setKeyPairsDispatcher(async (ctx, identifier) => {
-    const user = db
-      .prepare<unknown[], User>("SELECT * FROM users WHERE username = ?")
-      .get(identifier);
-    if (user == null) return [];
-
-    const rows = db
-      .prepare<unknown[], Key>("SELECT * FROM keys WHERE keys.user_id = ?")
-      .all(user.id);
-
-    const keys = Object.fromEntries(
-      rows.map((row) => [row.type, row]),
-    ) as Record<Key["type"], Key>;
-
-    const pairs: CryptoKeyPair[] = [];
-    for (const keyType of ["RSASSA-PKCS1-v1_5", "Ed25519"] as const) {
-      if (keys[keyType] == null) {
-        logger.debug(
-          "The user {identifier} does not have an {keyType} key; creating one...",
-          { identifier, keyType },
-        );
-        const { privateKey, publicKey } = await generateCryptoKeyPair(keyType);
-        db.prepare(
-          `
-            INSERT INTO keys (user_id, type, private_key, public_key)
-            VALUES (?, ?, ?, ?)
-          `,
-        ).run(
-          user.id,
-          keyType,
-          JSON.stringify(await exportJwk(privateKey)),
-          JSON.stringify(await exportJwk(publicKey)),
-        );
-        pairs.push({ privateKey, publicKey });
-      } else {
-        pairs.push({
-          privateKey: await importJwk(
-            JSON.parse(keys[keyType].private_key),
-            "private",
-          ),
-          publicKey: await importJwk(
-            JSON.parse(keys[keyType].public_key),
-            "public",
-          ),
-        });
-      }
-    }
-    return pairs;
-  });
-
-federation
-  .setInboxListeners("/users/{identifier}/inbox", "/inbox")
-  .on(Follow, async (ctx, follow) => {
-    if (follow.objectId == null) {
-      logger.debug("The follow object does not have an object: {follow}.", {
-        follow,
-      });
-      return;
-    }
-    const object = ctx.parseUri(follow.objectId);
-    if (object == null || object.type !== "actor") {
-      logger.debug("The follow object is not an actor: {follow}", {
-        follow,
-      });
-      return;
-    }
-    const follower = await follow.getActor();
-    if (follower?.id == null || follower.inboxId == null) {
-      logger.debug("The Follow object does not have an actor: {follow}", {
-        follow,
-      });
-      return;
-    }
-    const followingId = db
-      .prepare<unknown[], Actor>(
-        `
-          SELECT * FROM actors
-          JOIN users ON users.id = actors.user_id
-          WHERE users.username = ?
-        `,
-      )
-      .get(object.identifier)?.id;
-    if (followingId == null) {
-      logger.debug(
-        "Failed to find actor to follow on database: {object}",
-        { object },
-      );
-    }
-    // const followerId = db
-    //   .prepare<unknown[], Actor>(
-    //     `
-    //       -- Add a new follower actor record or update if it already exists
-    //       INSERT INTO actors (uri, handle, name, inbox_url, shared_inbox_url, url)
-    //       VALUES (?, ?, ?, ?, ?, ?)
-    //       ON CONFLICT (uri) DO UPDATE SET
-    //         handle = excluded.handle,
-    //         name = excluded.handle,
-    //         inbox_url = excluded.inbox_url,
-    //         shared_inbox_url = excluded.shared_inbox_url,
-    //         url = excluded.url
-    //       WHERE
-    //         actors.uri = excluded.uri
-    //       RETURNING *
-    //     `,
-    //   )
-    //   .get(
-    //     follower.id.href,
-    //     await getActorHandle(follower),
-    //     follower.name?.toString(),
-    //     follower.inboxId.href,
-    //     follower.endpoints?.sharedInbox?.href,
-    //     follower.url?.href,
-    //   )?.id;
-    const followerId = (await persistActor(follower))?.id;
-    db.prepare(
-      "INSERT INTO follows (following_id, follower_id) VALUES (?, ?)",
-    ).run(followingId, followerId);
-    const accept = new Accept({
-      actor: follow.objectId,
-      to: follow.actorId,
-      object: follow,
-    });
-    await ctx.sendActivity(object, follower, accept);
-  })
-  .on(Accept, async (ctx, accept) => {
-    const follow = await accept.getObject();
-    if (!(follow instanceof Follow)) return;
-    const following = await accept.getActor();
-    if (!isActor(following)) return;
-    const follower = follow.actorId;
-    if (follower == null) return;
-    const parsed = ctx.parseUri(follower);
-    if (parsed == null || parsed.type !== "actor") return;
-    const followingId = (await persistActor(following))?.id;
-    if (followingId == null) return;
-    db.prepare(
-      `
-        INSERT INTO follows (following_id, follower_id)
-        VALUES(
-          ?,
-          (
-            SELECT actors.id
-            FROM actors
-            JOIN users ON actors.user_id = users.id
-            WHERE users.username = ?
-          )
-        )
-      `,
-    ).run(followingId, parsed.identifier);
-  })
-  .on(Undo, async (ctx, undo) => {
-    const object = await undo.getObject();
-    // only handle undo of follow for now
-    if (!(object instanceof Follow)) return;
-    if (undo.actorId == null || object.objectId == null) return;
-    const parsed = ctx.parseUri(object.objectId);
-    console.debug(parsed);
-    if (parsed == null || parsed.type !== "actor") return;
-    db.prepare(
-      `
-        DELETE FROM follows
-        WHERE following_id = (
-          SELECT actors.id
-          FROM actors
-          JOIN users ON actors.user_id = users.id
-          WHERE users.username = ?
-        ) AND follower_id = (SELECT id FROM actors WHERE uri = ?)
-      `,
-    ).run(parsed.identifier, undo.actorId.href);
-    console.debug("7");
-  })
-  .on(Create, async (ctx, create) => {
-    const object = await create.getObject();
-    if (!(object instanceof Note)) return;
-    const actor = create.actorId;
-    if (actor == null) return;
-    const author = await object.getAttribution();
-    if (!isActor(author) || author.id?.href !== actor.href) return; // no boosts??
-    const actorId = (await persistActor(author))?.id;
-    if (actorId == null) return;
-    if (object.id == null) return;
-    const content = object.content?.toString();
-    db.prepare(
-      "INSERT INTO posts (uri, actor_id, content, url) VALUES (?, ?, ?, ?)",
-    ).run(object.id.href, actorId, content, object.url?.href);
-  });
-
-federation
-  .setFollowersDispatcher(
-    "/users/{identifier}/followers",
-    (ctx, identifier, cursor) => {
-      const followers = db
-        .prepare<unknown[], Actor>(
-          `
-            SELECT followers.*
-            FROM follows
-            JOIN actors AS followers ON follows.follower_id = followers.id
-            JOIN actors AS following ON follows.following_id = following.id
-            JOIN users ON users.id = following.user_id
-            WHERE users.username = ?
-            ORDER BY follows.created DESC
-          `,
-        )
-        .all(identifier);
-      const items: Recipient[] = followers.map((f) => ({
-        id: new URL(f.uri),
-        inboxId: new URL(f.inbox_url),
-        endPoints: f.shared_inbox_url == null
-          ? null
-          : { sharedInbox: new URL(f.shared_inbox_url) },
-      }));
-      return { items };
-    },
-  )
-  .setCounter((ctx, identifier) => {
-    const result = db
-      .prepare<unknown[], { cunt: number }>(
-        `
-          SELECT Count(*) AS cunt
-          FROM follows
-          JOIN actors ON actors.id = follows.following_id
-          JOIN users ON users.id = actors.user_id
-          WHERE users.username = ?
-        `,
-      )
-      .get(identifier);
-    return result == null ? 0 : result.cunt;
-  });
-
-federation.setObjectDispatcher(
-  Note,
-  "/users/{identifier}/posts/{id}",
-  (ctx, values) => {
-    const post = db
-      .prepare<unknown[], Post>(
-        `
-          SELECT posts.*
-          FROM posts
-          JOIN actors ON actors.id = posts.actor_id
-          JOIN users ON users.id = actors.user_id
-          WHERE users.username = ? AND posts.id = ?
-        `,
-      )
-      .get(values.identifier, values.id);
-    if (post == null) return null;
-    return new Note({
-      id: ctx.getObjectUri(Note, values),
-      attribution: ctx.getActorUri(values.identifier),
-      to: PUBLIC_COLLECTION,
-      cc: ctx.getFollowersUri(values.identifier),
-      content: post.content,
-      mediaType: "text/html",
-      published: Temporal.Instant.from(`${post.created.replace(" ", "T")}Z`),
-      url: ctx.getObjectUri(Note, values),
-    });
-  },
-);
-
-async function persistActor(actor: APActor): Promise<Actor | null> {
-  if (actor.id == null || actor.inboxId == null) {
-    logger.debug("Actor is missing required feilds: {actor}", { actor });
-    return null;
-  }
-  return (
-    db.prepare<unknown[], Actor>(
-      `
-        -- Add a new follower actor record or update if it already exists
-        INSERT INTO actors (uri, handle, name, inbox_url, shared_inbox_url, url)
-        VALUES (?, ?, ?, ?, ?, ?)
-        ON CONFLICT (uri) DO UPDATE SET
-          handle = excluded.handle,
-          name = excluded.handle,
-          inbox_url = excluded.inbox_url,
-          shared_inbox_url = excluded.shared_inbox_url,
-          url = excluded.url
-        WHERE
-          actors.uri = excluded.uri
-        RETURNING *
-      `,
-    )
-      .get(
-        actor.id.href,
-        await getActorHandle(actor),
-        actor.name?.toString(),
-        actor.inboxId.href,
-        actor.endpoints?.sharedInbox?.href,
-        actor.url?.href,
-      ) ?? null
-  );
-}
-
-export default federation;
