+++ 3o14/src/index.ts
@@ -1,7 +1,7 @@
 import { serve } from "@hono/node-server";
 import { behindProxy } from "x-forwarded-fetch";
 import app from "./routes/index.ts";
-import "./logging.ts";
+import "./config/logging.ts";
 
 serve(
   {
